fix(stagewise): split UI zoom from tab-content zoom hotkey handlers
Moved UI zoom (Cmd+=, Cmd+-, Cmd+0) to GlobalHotkeyBindings so it works
even when no content panel is open. Tab-content zoom stays in
BrowserTabHotkeys. Coordination via focusedPanel gating and listener
registration order.
